pyproject.toml
[tool.poetry]
name = "yards"
version = "0.1.1"
description = ""
authors = ["chanhakim <chanhakim17@gmail.com>", "jdnk <jadenisalive@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.19.0"
pyyaml = "^5.3.1"
progressbar2 = "^3.51.4"
pillow = "^7.2.0"
tqdm = "^4.48.0"
ray = "^0.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
yards = "yards._cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
